+++ .eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  extends: ['expo', 'prettier'],
-  plugins: ['prettier'],
-  rules: {
-    'prettier/prettier': 'error',
-  },
-};