+++ android/src/main/java/expo/modules/zebrascanner/ZebraConstants.kt
@@ -1,0 +1,5 @@
+package expo.modules.zebrascanner
+
+internal const val ACTION_BARCODE_SCANNED = "com.symbol.datawedge.ACTION_BARCODE_SCANNED"
+internal const val EVENT_BARCODE_SCANNED = "onBarcodeScanned"
+internal const val EVENT_CUSTOM_SCAN = "onCustomScan"