+++ src/internal/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_BARCODE_ACTION = "com.symbol.datawedge.ACTION_BARCODE_SCANNED";
+export const DATAWEDGE_API_ACTION = "com.symbol.datawedge.api.ACTION";